requirements.txt
SQLAlchemy==1.3.20
Flask-SQLAlchemy==2.4.4
graphene-sqlalchemy
Flask-GraphQL
logicbank
logicbankutils
requests